Крупный патч Lua и PPU
Note: * пересобрана структура проекта (все header файлы теперь в include/)
      * оптимизация PPU (на самом деле не совсем, но эффект тот же)
      * изменена логика обращений к Lua вызовам (вместо read/write
	PRG/CHR addr funcCall'ы)
      * небольшие изменения в lib.lua
      * исправлены тайминги ppu, теперь все должно отображаться
	корректно

+++ src/lua.cpp
@@ -10,7 +10,7 @@ enum VecType : int
 
 extern "C"
 {
-    void Cartridge_resize(void* instance, int vecType, size_t size) {
+    void Cartridge_resize(void* instance, u8 vecType, size_t size) {
         auto* cart = static_cast<Cartridge*>(instance);
 
         switch (vecType) {
